requirements.txt
flask==2.0.1
Werkzeug==2.0.1
requests==2.28.1
numpy
matplotlib
opencv-python
tensorflow==2.14.0
ultralytics
opencv-python
torch
torchvision
pyyaml
tensorflow
pillow
numpy
speciesnet
ultralytics
chromadb 
PySocks 
beautifulsoup4 
blinker 
deepface 
flask-cors 
gdown 
gunicorn 
itsdangerous  
lz4
ml-dtypes 
mtcnn 
retina-face 
soupsieve 
tensorboard
tf-keras
# faiss-cpu
faiss-gpu
